# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EduConnect API - connect students with tutors
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gamification/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [gamification]
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, allTime]
      responses:
        "200":
          description: Ranked users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
...
  /library:
    get:
      operationId: listLibraryBooks
      tags: [library]
      summary: List books with optional category and search filters
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of library books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryBook"
  /library/{id}:
    get:
      operationId: getLibraryBook
      tags: [library]
      summary: Get a single book with full details and AI summary
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Book detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryBookDetail"
...
  /lucky-royal/{id}:
    get:
      operationId: getLuckyRoyalRoom
      tags: [marketplace]
      summary: Get a single Lucky Royal room
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Room detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LuckyRoyalRoom"
  /lucky-royal/{id}/spin:
    post:
      operationId: spinLuckyRoyal
      tags: [marketplace]
      summary: Enter and spin a Lucky Royal room
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Spin result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LuckyRoyalResult"
        "400":
          description: Insufficient credits or room closed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                required: [error]
...
components:
  schemas:
...
    LeaderboardEntry:
      type: object
      properties:
        rank: { type: integer }
        userName: { type: string }
        avatarUrl: { type: string }
        points: { type: integer }
        levelLabel: { type: string }
        isMe: { type: boolean }
      required: [rank, userName, avatarUrl, points, levelLabel, isMe]
...
    LibraryBook:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        author: { type: string }
        category: { type: string }
        coverUrl: { type: string, nullable: true }
        shortDesc: { type: string }
        rating: { type: number }
        reviewCount: { type: integer }
        downloads: { type: integer }
        isFree: { type: boolean }
        priceCredits: { type: integer }
        pages: { type: integer }
        readingTimeMinutes: { type: integer }
        tags:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, title, author, category, shortDesc, rating, reviewCount, downloads, isFree, priceCredits, pages, readingTimeMinutes, tags, createdAt]
    LibraryBookDetail:
      allOf:
        - $ref: "#/components/schemas/LibraryBook"
        - type: object
          properties:
            description: { type: string }
            aiSummary: { type: string, nullable: true }
            contentPreview: { type: string, nullable: true }
          required: [description]
...
    LuckyRoyalRoom:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        creatorName: { type: string }
        creatorAvatar: { type: string, nullable: true }
        rewardTitle: { type: string }
        rewardImage: { type: string, nullable: true }
        rewardType: { type: string }
        entryCredits: { type: integer }
        maxParticipants: { type: integer }
        currentParticipants: { type: integer }
        endsAt: { type: string, format: date-time }
        status: { type: string }
        category: { type: string }
        isFeatured: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, title, description, creatorName, rewardTitle, rewardType, entryCredits, maxParticipants, currentParticipants, endsAt, status, category, isFeatured, createdAt]
    LuckyRoyalResult:
      type: object
      properties:
        resultType: { type: string }
        rewardLabel: { type: string }
        rewardAmount: { type: integer }
        newBalance: { type: integer }
        message: { type: string }
      required: [resultType, rewardLabel, rewardAmount, newBalance, message]
...